+++ lib/tio_popup/tio_popup_component/tio_popup_component.dart
@@ -227,7 +227,7 @@ RelativePosition _calcBestPosition({@required Rectangle<num> container,
     final overlapArea = overlapRect.width * overlapRect.height;
 
     if (overlapArea > bestOverlap) {
-      bestOverlap = overlapArea;
+      bestOverlap = overlapArea.toDouble();
       bestPosition = relativePosition;
     }
   }
